@@ -4,8 +4,8 @@ repository: https://github.com/enjoymoments/custom_view
 description: Package with widgets for build screens easly
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
-  flutter: 2.8.1
+  sdk: ">=2.16.2 <3.0.0"
+  flutter: 3.3.9
 
 dependencies:
   flutter:
