@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 extension Snackbar on BuildContext {
   void showSnackBar(String message) {
-    Scaffold.of(this)
+    ScaffoldMessenger.of(this)
       ..hideCurrentSnackBar()
       ..showSnackBar(
         SnackBar(content: Text(message)),
